pyproject.toml
[project]
  name            = "evfinder-api"
  version         = "0.1.0"
  description     = "Backend API services for the EV Finder."
  readme          = "README.md"
  requires-python = ">=3.14"
  dependencies    = [
    "fastapi>=0.128.0",
    "google-cloud-error-reporting>=1.14.0",
    "httpx[http2]>=0.28.1",
    "uvicorn[standard]>=0.40.0",
]

[dependency-groups]
  dev = [
    "faker>=40.1.0",
    "ipykernel>=7.1.0",
    "pre-commit>=4.5.1",
    "pytest>=9.0.2",
    "ruff>=0.14.11",
    "vcrpy>=8.1.1",
]
